==> notebooks/1.0-pj-extract-titanic-data.ipynb <==
!pip install python-dotenv

from dotenv import load_dotenv, find_dotenv

# find .env automatically by walking up dirs intil it's found
doenv_path = find_dotenv('*.env')
# load up the entries as environment variables
load_dotenv(doenv_path)

import os
KAGGLE_USERNAME = os.environ.get("KAGGLE_USERNAME")
print(KAGGLE_USERNAME)

import requests
from requests import session
import os
from dotenv import load_dotenv, find_dotenv

# payload for post
#payload = {
#    'action': 'login',
#    'username': os.environ.get("KAGGLE_USERNAME"),
#    'password': os.environ.get("KAGGLE_PASSWORD")
#}

# url for train file on Kaggle website
#url = 'https://www.kaggle.com/c/titanic/download/train.csv'

# setup session
#with session() as c:
    # post request
#    c.post('https://www.kaggle.com/account/login', data=payload)
    # get request
#    response = c.get(url)
    # print response text
#    print(response.text)

!ls -l ../data/raw
